fixed even more 64 bits warning
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1443 689a6050-402a-0410-94f2-e92a70836424

+++ parser/parse_privacy.h
@@ -45,7 +45,7 @@ enum privacy_value {
  * casting macro for accessing enumeration of priv-values
  */
 #define get_privacy_values(p_msg) \
-	((unsigned int)((p_msg)->privacy->parsed))
+	((unsigned int)(long)((p_msg)->privacy->parsed))
 
 
 /*
